MODWRKFLOW-57: Prevent double-posting of Workflow.
Utilize `RepositoryEventHandler` annotation to pre-process the Workflow before creating.
Detect if the workflow already exists and if it does then throw an error.

Implement the error handling functionality.
Make sure this functionality is consistently being handled for all defined exceptions.

Perform this check relative to the version tag string to allow for different versions of the same UUID.